いいね機能修正-1
