+++ sql/migration.sql
@@ -9,9 +9,9 @@ SET NAMES utf8mb4;
 DROP TABLE IF EXISTS `members`;
 CREATE TABLE `members` (
     `id` int NOT NULL AUTO_INCREMENT,
-    `name` varchar(255) DEFAULT NULL,
-    `email` varchar(255) DEFAULT NULL,
-    `password` varchar(255) DEFAULT NULL,
+    `name` varchar(255) NOT NULL,
+    `email` varchar(255) NOT NULL,
+    `password` varchar(255) NOT NULL,
     `icon` varchar(255) DEFAULT NULL,
     `created` timestamp NULL DEFAULT CURRENT_TIMESTAMP,
     `modified` timestamp NULL DEFAULT CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP,
@@ -22,7 +22,7 @@ CREATE TABLE `members` (
 DROP TABLE IF EXISTS `posts`;
 CREATE TABLE `posts` (
     `id` int NOT NULL AUTO_INCREMENT,
-    `member_id` int DEFAULT NULL,
+    `member_id` int NOT NULL,
     `message` text,
     `reply_post_id` int DEFAULT NULL,
     `created` timestamp NULL DEFAULT CURRENT_TIMESTAMP,
@@ -33,8 +33,8 @@ CREATE TABLE `posts` (
 DROP TABLE IF EXISTS `likes`;
 CREATE TABLE `likes` (
     `id` int NOT NULL AUTO_INCREMENT,
-    `post_id` int DEFAULT NULL,
-    `liker_id` int DEFAULT NULL,
+    `post_id` int NOT NULL,
+    `member_id` int NOT NULL,
     `created` timestamp NULL DEFAULT CURRENT_TIMESTAMP,
     `modified` timestamp NULL DEFAULT CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP,
     PRIMARY KEY (`id`)
